+++ src/components/donation/Steps/Submit/Crypto/index.ts
@@ -1,1 +1,1 @@
-export { default } from "./Crypto";
+export { default } from "./Checkout";
